commit transaction history payee

+++ Common/Models/Response/TransactionHistoryResponse.cs
@@ -22,7 +22,7 @@ namespace Common.Models.Response
         public string? TransactionFrom { get; set; }
         public string? TransactionTo { get; set; }
         public DateTime CreatedOn { get; set; }
-        public int Amount { get; set; }
+        public double Amount { get; set; }
         public string? Note { get; set; }
         public string? TransactionType { get; set; }
         public bool IsSelfTransfer { get; set; }
